Fix handlers persisting after file change
The other alternative to global handler storage is supposedly using the driver namespace, and I feel that's worse than globals so globals it is.
These shouldn't need to be updated from the remove() function as then there will be no callback functions being ran to check for the global handlers anyways.